requirement.txt
blinker==1.6.2
bson==0.5.10
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.2
Flask-JWT-Extended==4.5.2
fpdf==1.7.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
idna==3.4
im2pdf==0.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybase64==1.3.0
PyJWT==2.8.0
pymongo==4.4.1
PyPDF2==3.0.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7